requirements.txt
pyyaml
discord
python-dotenv